now just need some aesthetic adjustments & rest of the data
