--- src/routes/blog/sensitivity-analysis/lib/city_coords2.js
@@ -1,73 +1,0 @@
-export const cityCoordinates = {
-    'Albuquerque NM': { lat: '35.0844', long: '-106.6504', rq_comm: 0.6294742, rq_zip: 0.9072551, rq_hdbscan: 0.6624530 },
-    'Atlanta GA': { lat: '33.7490', long: '-84.3880', rq_comm: 0.5977573, rq_zip: 0.8477988, rq_hdbscan: 0.8525681 }
-};
-
-//     'Albuquerque NM': { lat: '35.0844', long: '-106.6504'},
-//     'Atlanta GA': { lat: '33.7490', long: '-84.3880'},
-//     'Austin TX': { lat: '30.2672', long: '-97.7431' },
-//     'Bakersfield CA': { lat: '35.3733', long: '-119.0187' },
-//     'Baltimore MD': { lat: '39.2904', long: '-76.6122' },
-//     'Boston MA': { lat: '42.3601', long: '-71.0589' },
-//     'Calgary AB': { lat: '51.0447', long: '-114.0719' },
-//     'Charlotte NC': { lat: '35.2271', long: '-80.8431' },
-//     'Chicago IL': { lat: '41.8781', long: '-87.6298' },
-//     'Cincinnati OH': { lat: '39.1031', long: '-84.5120' },
-//     'Cleveland OH': { lat: '41.4993', long: '-81.6944' },
-//     'Colorado Springs CO': { lat: '38.8339', long: '-104.8214' },
-//     'Columbus OH': { lat: '39.9612', long: '-82.9988' },
-//     'Dallas TX': { lat: '32.7767', long: '-96.7970' },
-//     'Denver CO': { lat: '39.7392', long: '-104.9903' },
-//     'Detroit MI': { lat: '42.3314', long: '-83.0458' },
-//     'Edmonton AB': { lat: '53.5444', long: '-113.4909' },
-//     'El Paso TX': { lat: '31.7619', long: '-106.4850' },
-//     'Fort Worth TX': { lat: '32.7555', long: '-97.3308' },
-//     'Fresno CA': { lat: '36.735', long: '-119.786' },
-//     'Halifax NS': { lat: '44.6488', long: '-63.5752' },
-//     'Honolulu HI': { lat: '21.3069', long: '-157.8583' },
-//     'Houston TX': { lat: '29.7604', long: '-95.3698' },
-//     'Indianapolis IN': { lat: '39.7684', long: '-86.1581' },
-//     'Jacksonville FL': { lat: '30.3322', long: '-81.6557' },
-//     'Kansas City MO': { lat: '39.0997', long: '-94.5786' },
-//     'Las Vegas NV': { lat: '36.1699', long: '-115.1398' },
-//     'London ON': { lat: '42.9849', long: '-81.2453' },
-//     'Los Angeles CA': { lat: '34.0522', long: '-118.2437' },
-//     'Louisville KY': { lat: '38.2529', long: '-85.7510' },
-//     'Memphis TN': { lat: '35.1495', long: '-90.0490' },
-//     'Miami FL': { lat: '25.7617', long: '-80.1918' },
-//     'Milwaukee WI': { lat: '43.0389', long: '-87.9065' },
-//     'Minneapolis MN': { lat: '44.9778', long: '-93.2650' },
-//     'Mississauga ON': { lat: '43.5890', long: '-79.6441' },
-//     'Montreal QC': { lat: '45.5017', long: '-73.5673' },
-//     'Nashville TN': { lat: '36.1627', long: '-86.7816' },
-//     'New Orleans LA': { lat: '29.9511', long: '-90.0715' },
-//     'New York NY': { lat: '40.727', long: '-73.9922' },
-//     'Oakland CA': { lat: '37.8044', long: '-122.2711' },
-//     'Oklahoma City OK': { lat: '35.4676', long: '-97.5164' },
-//     'Omaha NE': { lat: '41.2565', long: '-95.9345' },
-//     'Orlando FL': { lat: '28.5383', long: '-81.3792' },
-//     'Ottawa ON': { lat: '45.4215', long: '-75.6919' },
-//     'Philadelphia PA': { lat: '39.9526', long: '-75.1652' },
-//     'Phoenix AZ': { lat: '33.4484', long: '-112.0740' },
-//     'Pittsburgh PA': { lat: '40.4406', long: '-79.9959' },
-//     'Portland OR': { lat: '45.52', long: '-122.6750' },
-//     'Quebec QC': { lat: '46.8139', long: '-71.2080' },
-//     'Raleigh NC': { lat: '35.7796', long: '-78.6382' },
-//     'Sacramento CA': { lat: '38.5816', long: '-121.4944' },
-//     'Salt Lake City UT': { lat: '40.7608', long: '-111.8910' },
-//     'San Antonio TX': { lat: '29.4241', long: '-98.4936' },
-//     'San Diego CA': { lat: '32.7157', long: '-117.1611' },
-//     'San Francisco CA': { lat: '37.7921', long: '-122.4035' },
-//     'San Jose CA': { lat: '37.3355', long: '-121.8922' },
-//     'Seattle WA': { lat: '47.6062', long: '-122.3321' },
-//     'St. Louis MO': { lat: '38.6270', long: '-90.1994' },
-//     'Tampa FL': { lat: '27.9506', long: '-82.4572' },
-//     'Toronto ON': { lat: '43.653', long: '-79.386' },
-//     'Tucson AZ': { lat: '32.2226', long: '-110.9747' },
-//     'Tulsa OK': { lat: '36.1540', long: '-95.9928' },
-//     'Vancouver BC': { lat: '49.2827', long: '-123.1207' },
-//     'Washington DC': { lat: '38.8951', long: '-77.0369' },
-//     'Wichita KS': { lat: '37.6872', long: '-97.3301' },
-//     'Winnipeg MB': { lat: '49.8951', long: '-97.1384' }
-//   };
-
