--- src/routes/blog/sensitivity-analysis/lib/city_coords3.js
@@ -1,466 +1,0 @@
-export const cityCoordinates = 
-{
-    'Albuquerque NM': {
-        'lat': '35.0844',
-        'long': '-106.6504',
-        'rq_comm': 0.62947423,
-        'rq_zip': 0.907255071,
-        'rq_hdbscan': 0.662453049
-    },
-    'Atlanta GA': {
-        'lat': '33.7490',
-        'long': '-84.3880',
-        'rq_comm': 0.5977573,
-        'rq_zip': 0.847798765,
-        'rq_hdbscan': 0.852568124
-    },
-    'Austin TX': {
-        'lat': '30.2672',
-        'long': '-97.7431',
-        'rq_comm': 0.744620969,
-        'rq_zip': 0.775481836,
-        'rq_hdbscan': 0.726433542
-    },
-    'Bakersfield CA': {
-        'lat': '35.3733',
-        'long': '-119.0187',
-        'rq_comm': 1.183083598,
-        'rq_zip': 0.980774813,
-        'rq_hdbscan': 0.945859469
-    },
-    'Baltimore MD': {
-        'lat': '39.2904',
-        'long': '-76.6122',
-        'rq_comm': 0.69900637,
-        'rq_zip': 0.797767374,
-        'rq_hdbscan': 0.677724457
-    },
-    'Boston MA': {
-        'lat': '42.3601',
-        'long': '-71.0589',
-        'rq_comm': 0.781461902,
-        'rq_zip': 0.610299369,
-        'rq_hdbscan': 0.702302144
-    },
-    'Calgary AB': {
-        'lat': '51.0447',
-        'long': '-114.0719',
-        'rq_comm': 0.761676001,
-        'rq_zip': 0.849609754,
-        'rq_hdbscan': 0.851878221
-    },
-    'Charlotte NC': {
-        'lat': '35.2271',
-        'long': '-80.8431',
-        'rq_comm': 0.633754472,
-        'rq_zip': 0.847648601,
-        'rq_hdbscan': 0.745151012
-    },
-    'Chicago IL': {
-        'lat': '41.8781',
-        'long': '-87.6298',
-        'rq_comm': 0.701135982,
-        'rq_zip': 0.612941651,
-        'rq_hdbscan': 0.614576626
-    },
-    'Cincinnati OH': {
-        'lat': '39.1031',
-        'long': '-84.5120',
-        'rq_comm': 0.674160228,
-        'rq_zip': 0.75336364,
-        'rq_hdbscan': 0.644403997
-    },
-    'Cleveland OH': {
-        'lat': '41.4993',
-        'long': '-81.6944',
-        'rq_comm': 0.662719502,
-        'rq_zip': 0.827013579,
-        'rq_hdbscan': 0.781381482
-    },
-    'Colorado Springs CO': {
-        'lat': '38.8339',
-        'long': '-104.8214',
-        'rq_comm': 0.8699984,
-        'rq_zip': 0.874053603,
-        'rq_hdbscan': 0.873607568
-    },
-    'Columbus OH': {
-        'lat': '39.9612',
-        'long': '-82.9988',
-        'rq_comm': 0.452794829,
-        'rq_zip': 0.722441549,
-        'rq_hdbscan': 0.565491365
-    },
-    'Dallas TX': {
-        'lat': '32.7767',
-        'long': '-96.7970',
-        'rq_comm': 0.822155033,
-        'rq_zip': 0.762260817,
-        'rq_hdbscan': 0.702780863
-    },
-    'Denver CO': {
-        'lat': '39.7392',
-        'long': '-104.9903',
-        'rq_comm': 0.683393082,
-        'rq_zip': 0.708027139,
-        'rq_hdbscan': 0.66794486
-    },
-    'Detroit MI': {
-        'lat': '42.3314',
-        'long': '-83.0458',
-        'rq_comm': 0.724539133,
-        'rq_zip': 0.711286113,
-        'rq_hdbscan': 0.658492057
-    },
-    'Edmonton AB': {
-        'lat': '53.5444',
-        'long': '-113.4909',
-        'rq_comm': 0.761949745,
-        'rq_zip': 0.742759346,
-        'rq_hdbscan': 0.799277619
-    },
-    'El Paso TX': {
-        'lat': '31.7619',
-        'long': '-106.4850',
-        'rq_comm': 0.76850743,
-        'rq_zip': 1.142143702,
-        'rq_hdbscan': 0.965238639
-    },
-    'Fort Worth TX': {
-        'lat': '32.7555',
-        'long': '-97.3308',
-        'rq_comm': 0.771218721,
-        'rq_zip': 0.91381382,
-        'rq_hdbscan': 0.801332894
-    },
-    'Fresno CA': {
-        'lat': '36.735',
-        'long': '-119.786',
-        'rq_comm': 0.670720169,
-        'rq_zip': 0.776116103,
-        'rq_hdbscan': 0.718326524
-    },
-    'Halifax NS': {
-        'lat': '44.6488',
-        'long': '-63.5752',
-        'rq_comm': 0.702902857,
-        'rq_zip': 0.804715515,
-        'rq_hdbscan': 0.849515942
-    },
-    'Honolulu HI': {
-        'lat': '21.3069',
-        'long': '-157.8583',
-        'rq_comm': 0, // GET THIS NUMBER FROM BYEONGHWA!!!
-        'rq_zip': 0.919192243,
-        'rq_hdbscan': 0.75818201
-    },
-    'Houston TX': {
-        'lat': '29.7604',
-        'long': '-95.3698',
-        'rq_comm': 0.656492842,
-        'rq_zip': 0.780319331,
-        'rq_hdbscan': 0.639464348
-    },
-    'Indianapolis IN': {
-        'lat': '39.7684',
-        'long': '-86.1581',
-        'rq_comm': 0.720460159,
-        'rq_zip': 0.831113593,
-        'rq_hdbscan': 0.79368052
-    },
-    'Jacksonville FL': {
-        'lat': '30.3322',
-        'long': '-81.6557',
-        'rq_comm': 0.746731985,
-        'rq_zip': 0.959214628,
-        'rq_hdbscan': 0.70245259
-    },
-    'Kansas City MO': {
-        'lat': '39.0997',
-        'long': '-94.5786',
-        'rq_comm': 0.665938182,
-        'rq_zip': 0.800580199,
-        'rq_hdbscan': 0.738697644
-    },
-    'Las Vegas NV': {
-        'lat': '36.1699',
-        'long': '-115.1398',
-        'rq_comm': 0.910141813,
-        'rq_zip': 1.12152602,
-        'rq_hdbscan': 1.02636461
-    },
-    'London ON': {
-        'lat': '42.9849',
-        'long': '-81.2453',
-        'rq_comm': 0.779525691,
-        'rq_zip': 0.825072701,
-        'rq_hdbscan': 0.790509862
-    },
-    'Los Angeles CA': {
-        'lat': '34.0522',
-        'long': '-118.2437',
-        'rq_comm': 0.910935556,
-        'rq_zip': 0.867965999,
-        'rq_hdbscan': 0.83190507
-    },
-    'Louisville KY': {
-        'lat': '38.2529',
-        'long': '-85.7510',
-        'rq_comm': 0.602049871,
-        'rq_zip': 0.757390467,
-        'rq_hdbscan': 0.542219287
-    },
-    'Memphis TN': {
-        'lat': '35.1495',
-        'long': '-90.0490',
-        'rq_comm': 0.629771677,
-        'rq_zip': 0.821322341,
-        'rq_hdbscan': 0.739803599
-    },
-    'Miami FL': {
-        'lat': '25.7617',
-        'long': '-80.1918',
-        'rq_comm': 0.928545334,
-        'rq_zip': 0.910486469,
-        'rq_hdbscan': 0.920359628
-    },
-    'Milwaukee WI': {
-        'lat': '43.0389',
-        'long': '-87.9065',
-        'rq_comm': 0.905307199,
-        'rq_zip': 0.98846817,
-        'rq_hdbscan': 0.857170402
-    },
-    'Minneapolis MN': {
-        'lat': '44.9778',
-        'long': '-93.2650',
-        'rq_comm': 0.444687606,
-        'rq_zip': 0.547229576,
-        'rq_hdbscan': 0.555528988
-    },
-    'Mississauga ON': {
-        'lat': '43.5890',
-        'long': '-79.6441',
-        'rq_comm': 0.835948006,
-        'rq_zip': 0.922684423,
-        'rq_hdbscan': 0.911440826
-    },
-    'Montreal QC': {
-        'lat': '45.5017',
-        'long': '-73.5673',
-        'rq_comm': 0.801717767,
-        'rq_zip': 0.860440604,
-        'rq_hdbscan': 0.669858491
-    },
-    'Nashville TN': {
-        'lat': '36.1627',
-        'long': '-86.7816',
-        'rq_comm': 0.808130952,
-        'rq_zip': 0.915306983,
-        'rq_hdbscan': 0.806812321
-    },
-    'New Orleans LA': {
-        'lat': '29.9511',
-        'long': '-90.0715',
-        'rq_comm': 0.559795411,
-        'rq_zip': 0.79355619,
-        'rq_hdbscan': 0.744811405
-    },
-    'New York NY': {
-        'lat': '40.727',
-        'long': '-73.9922',
-        'rq_comm': 0.76003437,
-        'rq_zip': 0.666758251,
-        'rq_hdbscan': 0.658605539
-    },
-    'Oakland CA': {
-        'lat': '37.8044',
-        'long': '-122.2711',
-        'rq_comm': 0.795924619,
-        'rq_zip': 0.840545872,
-        'rq_hdbscan': 0.743834817
-    },
-    'Oklahoma City OK': {
-        'lat': '35.4676',
-        'long': '-97.5164',
-        'rq_comm': 0.815284979,
-        'rq_zip': 0.93069373,
-        'rq_hdbscan': 0.929554482
-    },
-    'Omaha NE': {
-        'lat': '41.2565',
-        'long': '-95.9345',
-        'rq_comm': 0.576850061,
-        'rq_zip': 0.796789214,
-        'rq_hdbscan': 0.880426022
-    },
-    'Orlando FL': {
-        'lat': '28.5383',
-        'long': '-81.3792',
-        'rq_comm': 0.815668633,
-        'rq_zip': 0.918897402,
-        'rq_hdbscan': 0.791095319
-    },
-    'Ottawa ON': {
-        'lat': '45.4215',
-        'long': '-75.6919',
-        'rq_comm': 0.821161689,
-        'rq_zip': 0.83629477,
-        'rq_hdbscan': 0.822822024
-    },
-    'Philadelphia PA': {
-        'lat': '39.9526',
-        'long': '-75.1652',
-        'rq_comm': 0.637987897,
-        'rq_zip': 0.688932126,
-        'rq_hdbscan': 0.672484329
-    },    
-    'Phoenix AZ': {
-        'lat': '33.4484',
-        'long': '-112.0740',
-        'rq_comm': 0.704185622,
-        'rq_zip': 0.926850404,
-        'rq_hdbscan': 0.891630832
-    },
-    'Pittsburgh PA': {
-        'lat': '40.4406',
-        'long': '-79.9959',
-        'rq_comm': 0.67390962,
-        'rq_zip': 0.774942289,
-        'rq_hdbscan': 0.662843889
-    },
-    'Portland OR': {
-        'lat': '45.52',
-        'long': '-122.6750',
-        'rq_comm': 0.655381107,
-        'rq_zip': 0.51076172,
-        'rq_hdbscan': 0.607872146
-    },
-    'Quebec QC': {
-        'lat': '46.8139',
-        'long': '-71.2080',
-        'rq_comm': 0.749379558,
-        'rq_zip': 0.834849304,
-        'rq_hdbscan': 0.870927314
-    },
-    'Raleigh NC': {
-        'lat': '35.7796',
-        'long': '-78.6382',
-        'rq_comm': 0.618937997,
-        'rq_zip': 0.720619118,
-        'rq_hdbscan': 0.632144242
-    },
-    'Sacramento CA': {
-        'lat': '38.5816',
-        'long': '-121.4944',
-        'rq_comm': 0.820812766,
-        'rq_zip': 0.719509836,
-        'rq_hdbscan': 0.657146144
-    },
-    'Salt Lake City UT': {
-        'lat': '40.7608',
-        'long': '-111.8910',
-        'rq_comm': 0.88935258,
-        'rq_zip': 0.966512792,
-        'rq_hdbscan': 0.819177751
-    },
-    'San Antonio TX': {
-        'lat': '29.4241',
-        'long': '-98.4936',
-        'rq_comm': 0.687035594,
-        'rq_zip': 0.823519981,
-        'rq_hdbscan': 0.758478141
-    },
-    'San Diego CA': {
-        'lat': '32.7157',
-        'long': '-117.1611',
-        'rq_comm': 0.893547167,
-        'rq_zip': 0.976713504,
-        'rq_hdbscan': 0.796466854
-    },
-    'San Francisco CA': {
-        'lat': '37.7921',
-        'long': '-122.4035',
-        'rq_comm': 0.754418786,
-        'rq_zip': 0.60378238,
-        'rq_hdbscan': 0.673164327
-    },
-    'San Jose CA': {
-        'lat': '37.3355',
-        'long': '-121.8922',
-        'rq_comm': 1.075596962,
-        'rq_zip': 0.987428956,
-        'rq_hdbscan': 0.96491832
-    },
-    'Seattle WA': {
-        'lat': '47.6062',
-        'long': '-122.3321',
-        'rq_comm': 0.623177245,
-        'rq_zip': 0.6693637,
-        'rq_hdbscan': 0.565072365
-    },
-    'St. Louis MO': {
-        'lat': '38.6270',
-        'long': '-90.1994',
-        'rq_comm': 0.545597389,
-        'rq_zip': 0.588348265,
-        'rq_hdbscan': 0.531589
-    },
-    'Tampa FL': {
-        'lat': '27.9506',
-        'long': '-82.4572',
-        'rq_comm': 0.808130952,
-        'rq_zip': 0.915306983,
-        'rq_hdbscan': 0.806812321
-    },
-    'Toronto ON': {
-        'lat': '43.653',
-        'long': '-79.386',
-        'rq_comm': 0.732199918,
-        'rq_zip': 0.749099858,
-        'rq_hdbscan': 0.697519534
-    },
-    'Tucson AZ': {
-        'lat': '32.2226',
-        'long': '-110.9747',
-        'rq_comm': 0.803250237,
-        'rq_zip': 0.934612134,
-        'rq_hdbscan': 0.911952892
-    },
-    'Tulsa OK': {
-        'lat': '36.1540',
-        'long': '-95.9928',
-        'rq_comm': 0.685573391,
-        'rq_zip': 0.834301182,
-        'rq_hdbscan': 0.821499105
-    },
-    'Vancouver BC': {
-        'lat': '49.2827',
-        'long': '-123.1207',
-        'rq_comm': 0.899309392,
-        'rq_zip': 0.790859744,
-        'rq_hdbscan': 0.852328258
-    },
-    'Washington DC': {
-        'lat': '38.8951',
-        'long': '-77.0369',
-        'rq_comm': 0.71649651,
-        'rq_zip': 0.705903136,
-        'rq_hdbscan': 0.685057037
-    },
-    'Wichita KS': {
-        'lat': '37.6872',
-        'long': '-97.3301',
-        'rq_comm': 0.952408189,
-        'rq_zip': 0.870307538,
-        'rq_hdbscan': 0.890966784
-    },
-    'Winnipeg MB': {
-        'lat': '49.8951',
-        'long': '-97.1384',
-        'rq_comm': 0.5977573,
-        'rq_zip': 0.847798765,
-        'rq_hdbscan': 0.852568124
-    }
-}
-
